drm/msm/sde: Avoid dmesg flooding when DC dimming is enabled
For whatever reason, when DC dimming is enabled this logger has a
meltdown and repeatedly spams dmesg. I'm sure there's a better way
to handle this, but this is OnePlus code, we'll let them deal with
it. In the meantime, silence this error since nothing appears to
actually be broken.

Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>